Crate pallet_bridge

Bridge pallet

This pallet implements a bridge between Chainbridge and Centrifuge Chain.

Config

Call

Pallet

Overview

This pallet is used for bridging chains.

Terminology

Usage

Interface

Supported Origins

Signed origin is valid.

Types

BridgeOrigin - Specifies the origin check provided by the chainbridge for calls that can only be called by the chainbridge pallet. AdminOrigin - Admin user authorized to modify NativeTokenTransferFee and NftTokenTransferFee values. Currency - Currency as viewed from this pallet. Event - Type for events triggered by this pallet. NativeTokenId - Identifier of the native token. NativeTokenTransferFee - Additional fee charged for transfering native tokens. NftTokenTransferFee - Additional fee charged when moving NFTs to target chains. WeightInfo - Weight information for extrinsics in this pallet.

Events

Remark - Event triggered a remark proposal is approved.

Errors

InvalidTransfer - Invalid transfer. InsufficientBalance - Not enough resources/assets for performing a transfer. TotalAmountOverflow - Total amount to be transfered overflows balance type size.

Dispatchable Functions

Callable functions (or extrinsics), also considered as transactions, materialize the pallet contract. Here’s the callable functions implemented in this module:

[receive_nonfungible] [remark] [transfer] [transfer_asset] [transfer_native] [set_native_token_transfer_fee] [set_nft_token_transfer_fee]

Public Functions

Genesis Configuration

This pallet depends on the GenesisConfig. The following fields are added to the genesis configuration, that are not associated with specific storage values: chains: Vec<u8> - List of chains. relayers: Vec<T::AccountId>- List of relayers. resources: Vec<(ResourceId, Vec<u8>)> - List of resources (or assets). threshold: u32 - Threshold value.

Related Pallets

This pallet is tightly coupled to the following pallets:

Substrate FRAME’s balances pallet.

Centrifuge Chain chainbrige pallet.

Centrifuge Chain fees pallet.

Centrifuge Chain nft pallet.

Centrifuge Chain registry pallet.

References

Substrate FRAME v2 attribute macros.

Credits

The Centrifugians Tribe contributors@centrifuge.io

License

GNU General Public License, Version 3, 29 June 2007 https://www.gnu.org/licenses/gpl-3.0.html

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

traits

Traits used by bridge pallet