Trait pallet_bridge::pallet::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type RelayMembers: SortedMembers<Self::AccountId>;
    type UpdateTokenOrigin: EnsureOrigin<Self::Origin>;
    type UpdateChainOrigin: EnsureOrigin<Self::Origin>;
    type CapOrigin: EnsureOrigin<Self::Origin>;
    type RootOperatorAccountId: Get<Self::AccountId>;
    type Assets: Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance>;
    type GiftAccount: Get<Self::AccountId>;
    type GiftConvert: BalanceConversion<Balance, CurrencyId, Balance>;
    type NativeCurrencyId: Get<<<Self as Config>::Assets as Inspect<<Self as Config>::AccountId>>::AssetId>;
    type ExistentialDeposit: Get<Balance>;
    type ChainId: Get<ChainId>;
    type PalletId: Get<PalletId>;
    type ProposalLifetime: Get<Self::BlockNumber>;
    type ThresholdPercentage: Get<u32>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type RelayMembers: SortedMembers<Self::AccountId>

Relay members has permission to materialize the assets

type UpdateTokenOrigin: EnsureOrigin<Self::Origin>

The origin which can update bridged token

type UpdateChainOrigin: EnsureOrigin<Self::Origin>

The origin which can update bridged chain

type CapOrigin: EnsureOrigin<Self::Origin>

The origin which can clean accumulated cap value

type RootOperatorAccountId: Get<Self::AccountId>

The root operator account id

type Assets: Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance>

Assets for teleport/materialize assets to/from bridge pallet

type GiftAccount: Get<Self::AccountId>

An account to pay the bonus

type GiftConvert: BalanceConversion<Balance, CurrencyId, Balance>

A bonus amount converter

type NativeCurrencyId: Get<<<Self as Config>::Assets as Inspect<<Self as Config>::AccountId>>::AssetId>

Currency id of the native token

type ExistentialDeposit: Get<Balance>

The essential balance for an existed account

type ChainId: Get<ChainId>

The identifier for this chain. This must be unique and must not collide with existing IDs within a set of bridged chains.

type PalletId: Get<PalletId>

The bridge’s pallet id, keep all teleported assets.

type ProposalLifetime: Get<Self::BlockNumber>

Each proposal can live up to [ProposalLifetime] blocks

type ThresholdPercentage: Get<u32>

The threshold percentage of relayers required to approve a proposal

type WeightInfo: WeightInfo

Information on runtime weights.