Trait pallet_bridge::pallet::Config

pub trait Config: Config + Config {
    type BridgePalletId: Get<PalletId>;
    type BridgeOrigin: EnsureOrigin<<Self as Config>::RuntimeOrigin, Success = <Self as Config>::AccountId>;
    type Fees: Fees<AccountId = <Self as Config>::AccountId, Balance = <<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type Currency: Currency<<Self as Config>::AccountId>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type NativeTokenId: Get<ResourceId>;
    type NativeTokenTransferFeeKey: Get<<Self::Fees as Fees>::FeeKey>;
    type WeightInfo: WeightInfo;
}

Bridge pallet’s configuration trait.

Associated types and constants are declared in this trait. If the pallet depends on other super-traits, the latter must be added to this trait, Note that [frame_system::Config] must always be included.

Required Associated Types

type BridgePalletId: Get<PalletId>

Pallet identifier.

The module identifier may be of the form PalletId(*b"c/bridge") (a string of eight characters) and set using the parameter_types macro in one of the runtimes (see runtime folder).

type BridgeOrigin: EnsureOrigin<<Self as Config>::RuntimeOrigin, Success = <Self as Config>::AccountId>

Specifies the origin check provided by the chainbridge for calls that can only be called by the chainbridge pallet.

type Fees: Fees<AccountId = <Self as Config>::AccountId, Balance = <<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Entity used to pay fees

type Currency: Currency<<Self as Config>::AccountId>

Currency as viewed from this pallet

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Associated type for Event enum

type NativeTokenId: Get<ResourceId>

type NativeTokenTransferFeeKey: Get<<Self::Fees as Fees>::FeeKey>

Key used to retrieve the fee that are charged when transferring native tokens to target chains.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet